requirements.txt
google-cloud-aiplatform>=1.36.0
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
pytest>=7.4.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
python-jose>=3.3.0
aiohttp>=3.8.0 